+++ node/src/components/rest_server.rs
@@ -210,6 +210,8 @@ mod tests {
         let actual_schema = serde_json::to_string_pretty(&actual_schema).unwrap();
         let actual_schema: Value = serde_json::from_str(&actual_schema).unwrap();
 
+        // println!("{}", actual_schema);
+
         assert_json_eq!(actual_schema, expected_schema);
     }