+++ types/src/cl_value.rs
@@ -13,7 +13,7 @@ use serde_json::Value;
 
 use crate::{
     bytesrepr::{self, Bytes, FromBytes, ToBytes, U32_SERIALIZED_LENGTH},
-    checksummed_hex, CLType, CLTyped,
+    CLType, CLTyped,
 };
 
 mod jsonrepr;
@@ -203,7 +203,7 @@ impl Serialize for CLValue {
         if serializer.is_human_readable() {
             CLValueJson {
                 cl_type: self.cl_type.clone(),
-                bytes: checksummed_hex::encode(&self.bytes),
+                bytes: base16::encode_lower(&self.bytes),
                 parsed: jsonrepr::cl_value_to_json(self),
             }
             .serialize(serializer)
@@ -219,7 +219,7 @@ impl<'de> Deserialize<'de> for CLValue {
             let json = CLValueJson::deserialize(deserializer)?;
             (
                 json.cl_type.clone(),
-                checksummed_hex::decode(&json.bytes).map_err(D::Error::custom)?,
+                base16::decode(&json.bytes).map_err(D::Error::custom)?,
             )
         } else {
             <(CLType, Vec<u8>)>::deserialize(deserializer)?