+++ types/src/contract_wasm.rs
@@ -15,7 +15,7 @@ use crate::{
     account,
     account::TryFromSliceForAccountHashError,
     bytesrepr::{Bytes, Error, FromBytes, ToBytes},
-    checksummed_hex, uref, CLType, CLTyped, HashAddr,
+    uref, CLType, CLTyped, HashAddr,
 };
 
 const CONTRACT_WASM_MAX_DISPLAY_LEN: usize = 16;
@@ -105,7 +105,7 @@ impl ContractWasmHash {
 
     /// Formats the `ContractWasmHash` for users getting and putting.
     pub fn to_formatted_string(self) -> String {
-        format!("{}{}", WASM_STRING_PREFIX, checksummed_hex::encode(&self.0),)
+        format!("{}{}", WASM_STRING_PREFIX, base16::encode_lower(&self.0),)
     }
 
     /// Parses a string formatted as per `Self::to_formatted_string()` into a
@@ -114,20 +114,20 @@ impl ContractWasmHash {
         let remainder = input
             .strip_prefix(WASM_STRING_PREFIX)
             .ok_or(FromStrError::InvalidPrefix)?;
-        let bytes = HashAddr::try_from(checksummed_hex::decode(remainder)?.as_ref())?;
+        let bytes = HashAddr::try_from(base16::decode(remainder)?.as_ref())?;
         Ok(ContractWasmHash(bytes))
     }
 }
 
 impl Display for ContractWasmHash {
     fn fmt(&self, f: &mut Formatter<'_>) -> core::fmt::Result {
-        write!(f, "{}", checksummed_hex::encode(&self.0))
+        write!(f, "{}", base16::encode_lower(&self.0))
     }
 }
 
 impl Debug for ContractWasmHash {
     fn fmt(&self, f: &mut Formatter) -> core::fmt::Result {
-        write!(f, "ContractWasmHash({})", checksummed_hex::encode(&self.0))
+        write!(f, "ContractWasmHash({})", base16::encode_lower(&self.0))
     }
 }