+++ types/src/transfer.rs
@@ -21,7 +21,7 @@ use serde::{de::Error as SerdeError, Deserialize, Deserializer, Serialize, Seria
 use crate::{
     account::AccountHash,
     bytesrepr::{self, FromBytes, ToBytes},
-    checksummed_hex, CLType, CLTyped, URef, U512,
+    CLType, CLTyped, URef, U512,
 };
 
 /// The length of a deploy hash.
@@ -92,7 +92,7 @@ impl FromBytes for DeployHash {
 impl Serialize for DeployHash {
     fn serialize<S: Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
         if serializer.is_human_readable() {
-            checksummed_hex::encode(&self.0).serialize(serializer)
+            base16::encode_lower(&self.0).serialize(serializer)
         } else {
             self.0.serialize(serializer)
         }
@@ -103,8 +103,7 @@ impl<'de> Deserialize<'de> for DeployHash {
     fn deserialize<D: Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
         let bytes = if deserializer.is_human_readable() {
             let hex_string = String::deserialize(deserializer)?;
-            let vec_bytes =
-                checksummed_hex::decode(hex_string.as_bytes()).map_err(SerdeError::custom)?;
+            let vec_bytes = base16::decode(hex_string.as_bytes()).map_err(SerdeError::custom)?;
             <[u8; DEPLOY_HASH_LENGTH]>::try_from(vec_bytes.as_ref()).map_err(SerdeError::custom)?
         } else {
             <[u8; DEPLOY_HASH_LENGTH]>::deserialize(deserializer)?
@@ -294,7 +293,7 @@ impl TransferAddr {
         format!(
             "{}{}",
             TRANSFER_ADDR_FORMATTED_STRING_PREFIX,
-            checksummed_hex::encode(&self.0),
+            base16::encode_lower(&self.0),
         )
     }
 
@@ -303,8 +302,7 @@ impl TransferAddr {
         let remainder = input
             .strip_prefix(TRANSFER_ADDR_FORMATTED_STRING_PREFIX)
             .ok_or(FromStrError::InvalidPrefix)?;
-        let bytes =
-            <[u8; TRANSFER_ADDR_LENGTH]>::try_from(checksummed_hex::decode(remainder)?.as_ref())?;
+        let bytes = <[u8; TRANSFER_ADDR_LENGTH]>::try_from(base16::decode(remainder)?.as_ref())?;
         Ok(TransferAddr(bytes))
     }
 }
@@ -318,8 +316,7 @@ impl JsonSchema for TransferAddr {
     fn json_schema(gen: &mut SchemaGenerator) -> Schema {
         let schema = gen.subschema_for::<String>();
         let mut schema_object = schema.into_object();
-        schema_object.metadata().description =
-            Some("Checksummed hex-encoded transfer address.".to_string());
+        schema_object.metadata().description = Some("Hex-encoded transfer address.".to_string());
         schema_object.into()
     }
 }
@@ -346,35 +343,15 @@ impl<'de> Deserialize<'de> for TransferAddr {
     }
 }
 
-// impl TryFrom<&[u8]> for AccountHash {
-//     type Error = TryFromSliceForAccountHashError;
-//
-//     fn try_from(bytes: &[u8]) -> Result<Self, TryFromSliceForAccountHashError> {
-//         [u8; TRANSFER_ADDR_LENGTH]::try_from(bytes)
-//             .map(AccountHash::new)
-//             .map_err(|_| TryFromSliceForAccountHashError(()))
-//     }
-// }
-//
-// impl TryFrom<&alloc::vec::Vec<u8>> for AccountHash {
-//     type Error = TryFromSliceForAccountHashError;
-//
-//     fn try_from(bytes: &Vec<u8>) -> Result<Self, Self::Error> {
-//         [u8; TRANSFER_ADDR_LENGTH]::try_from(bytes as &[u8])
-//             .map(AccountHash::new)
-//             .map_err(|_| TryFromSliceForAccountHashError(()))
-//     }
-// }
-
 impl Display for TransferAddr {
     fn fmt(&self, f: &mut Formatter<'_>) -> core::fmt::Result {
-        write!(f, "{}", checksummed_hex::encode(&self.0))
+        write!(f, "{}", base16::encode_lower(&self.0))
     }
 }
 
 impl Debug for TransferAddr {
     fn fmt(&self, f: &mut Formatter) -> core::fmt::Result {
-        write!(f, "TransferAddr({})", checksummed_hex::encode(&self.0))
+        write!(f, "TransferAddr({})", base16::encode_lower(&self.0))
     }
 }